Rename RecordGroupSerdes to DeserializedRecordGroup.
