+++ components/core/src/reducer/DeserializedRecordGroup.hpp
